improved examples
I made the first example one in which the distinction between getting
up and waking up is important, and I made the second example one in
which the word “perhaps” actually weakens the point instead of giving a
numeric fuzz factor. These are not super captivating examples, but
hopefully illustrate my problem with the prior examples.