GH-15: flag words as rare instead of bad in NeoVim
This change marks all words in all dictionaries as rare words so that
NeoVim highlights them as SpellRare instead of SpellBad.  This lets you
distinguish between wordy writing and normal misspellings, respectively.

In order for this to work, your NeoVim needs to be patched as follows:

  https://github.com/neovim/neovim/pull/2456

  https://groups.google.com/forum/#!topic/vim_dev/rPWOoR3ZgSA

See `:help spell-wordlist-format` for details about dictionary markings.