Set spelllang correctly
Full paths are not allowed in spelllang, only the lang names.

In the ./spell/ directory the files are named using the 'lang' which in
the case of wordy is the dict name.

This patch drops the actual lang from the .spl path, and only passes
the dict names to spelllang.

NOTE: I am not a vimscript expert, this solution works for me when
loading vim-wordy using pathogen. Closes issue #33